+++ lib/modules/sudo/set_user_as_sudoer.sh
@@ -1,0 +1,6 @@
+set_user_as_sudoer () {
+    local user=$1
+    local sudoers_file="/etc/sudoers.d/admin"
+
+    sudo echo "$user ALL=(ALL) NOPASSWD:ALL" > $sudoers_file
+}