+++ lib/modules/users/ensure_admin_sudo.sh
@@ -1,0 +1,12 @@
+ensure_admin_sudo () {
+    local user=$1
+    local sudoers_file="/etc/sudoers.d/admin"
+
+    if $(is_user_sudoer $user); then
+        print ${SYMBOL_OK} "Admin user $user is a sudoer"
+    else
+        print "Setting $user as a sudoer..."
+        set_user_as_sudoer $user
+        print ${SYMBOL_OK} "Admin user $user is now a sudoer."
+    fi
+}